# Parse the instrument tree in the base instrument