--- a/Framework/DataHandling/src/LoadInstrument.cpp
+++ b/Framework/DataHandling/src/LoadInstrument.cpp
@@ -185,7 +185,7 @@ void LoadInstrument::exec() {
       // instrument. As a consequence less time is spent and less memory is
       // used. Note that this is only possible since the tree in `instrument`
       // will not be modified once we add it to the IDS.
-      instrument->parseTree();
+      instrument->parseTreeAndCacheBeamline();
       // Add to data service for later retrieval
       InstrumentDataService::Instance().add(instrumentNameMangled, instrument);
     }