--- a/Framework/Geometry/inc/MantidGeometry/Instrument.h
+++ b/Framework/Geometry/inc/MantidGeometry/Instrument.h
@@ -19,11 +19,9 @@ namespace Mantid {
 /// Typedef of a map from detector ID to detector shared pointer.
 typedef std::map<detid_t, Geometry::IDetector_const_sptr> detid2det_map;
 
-namespace Beamline {
+namespace Geometry {
 class ComponentInfo;
 class DetectorInfo;
-}
-namespace Geometry {
 class XMLInstrumentParameter;
 class ParameterMap;
 class ReferenceFrame;
@@ -249,6 +247,10 @@ public:
   /// Add a component to the instrument
   virtual int add(IComponent *component) override;
 
+  void parseTree();
+  const ComponentInfo *componentInfo() const;
+  const DetectorInfo *detectorInfo() const;
+
 private:
   /// Save information about a set of detectors to Nexus
   void saveDetectorSetInfoToNexus(::NeXus::File *file,
@@ -327,13 +329,11 @@ private:
   /// Pointer to the reference frame object.
   boost::shared_ptr<ReferenceFrame> m_referenceFrame;
 
-  /// Pointer to the DetectorInfo object. NULL unless the instrument is
-  /// associated with an ExperimentInfo object.
-  boost::shared_ptr<const Beamline::DetectorInfo> m_detectorInfo{nullptr};
+  /// Pointer to the DetectorInfo object. May be NULL.
+  boost::shared_ptr<const DetectorInfo> m_detectorInfo{nullptr};
 
-  /// Pointer to the ComponentInfo object. NULL unless the instrument is
-  /// associated with an ExperimentInfo object.
-  boost::shared_ptr<const Beamline::ComponentInfo> m_componentInfo{nullptr};
+  /// Pointer to the ComponentInfo object. May be NULL.
+  boost::shared_ptr<const ComponentInfo> m_componentInfo{nullptr};
 
   /// Flag - is this the physical rather than neutronic instrument
   bool m_isPhysicalInstrument{false};