--- a/Framework/Geometry/inc/MantidGeometry/Instrument.h
+++ b/Framework/Geometry/inc/MantidGeometry/Instrument.h
@@ -267,9 +267,6 @@ private:
   /// monitor flags.
   std::vector<std::tuple<detid_t, IDetector_const_sptr, bool>> m_detectorCache;
 
-  /// Mappings to obtain component index
-  boost::shared_ptr<const std::vector<Geometry::ComponentID>> m_componentCache;
-
   /// Purpose to hold copy of source component. For now assumed to be just one
   /// component
   const IComponent *m_sourceCache;