--- a/Framework/Geometry/inc/MantidGeometry/Instrument/DetectorInfo.h
+++ b/Framework/Geometry/inc/MantidGeometry/Instrument/DetectorInfo.h
@@ -23,7 +23,6 @@ class SpectrumInfo;
 namespace Geometry {
 class IDetector;
 class Instrument;
-class ParameterMap;
 
 /** Geometry::DetectorInfo is an intermediate step towards a DetectorInfo that
   is part of Instrument-2.0. The aim is to provide a nearly identical interface
@@ -128,7 +127,7 @@ public:
   void merge(const DetectorInfo &other);
 
   friend class API::SpectrumInfo;
-  friend class ParameterMap;
+  friend class Instrument;
 
 private:
   const Geometry::IDetector &getDetector(const size_t index) const;