--- a/Framework/Geometry/src/Instrument/ParameterMap.cpp
+++ b/Framework/Geometry/src/Instrument/ParameterMap.cpp
@@ -6,8 +6,6 @@
 #include "MantidGeometry/Objects/BoundingBox.h"
 #include "MantidGeometry/IDetector.h"
 #include "MantidKernel/Cache.h"
-#include "MantidBeamline/ComponentInfo.h"
-#include "MantidBeamline/DetectorInfo.h"
 #include "MantidKernel/MultiThreaded.h"
 #include "MantidGeometry/Instrument.h"
 #include "MantidGeometry/Instrument/ParameterFactory.h"
@@ -77,11 +75,9 @@ ParameterMap::ParameterMap(const ParameterMap &other)
           Kernel::make_unique<Kernel::Cache<const ComponentID, BoundingBox>>(
               *other.m_boundingBoxMap)),
       m_instrument(other.m_instrument) {
-  if (!other.m_detectorInfo)
-    return;
-  m_detectorInfo = Kernel::make_unique<DetectorInfo>(*other.m_detectorInfo);
-  m_componentInfo = Kernel::make_unique<ComponentInfo>(*other.m_componentInfo);
-  relinkWrappers();
+  if (m_instrument)
+    std::tie(m_componentInfo, m_detectorInfo) =
+        m_instrument->makeBeamline(*this, &other);
 }
 
 // Defined as default in source for forward declaration with std::unique_ptr.
@@ -1235,30 +1231,7 @@ void ParameterMap::setInstrument(const Instrument *instrument) {
     throw std::logic_error("ParameterMap::setInstrument must be called with "
                            "base instrument, not a parametrized instrument");
   m_instrument = instrument;
-
-  if (empty() && instrument->componentInfo()) {
-    m_detectorInfo =
-        Kernel::make_unique<DetectorInfo>(*instrument->detectorInfo());
-    m_componentInfo =
-        Kernel::make_unique<ComponentInfo>(*instrument->componentInfo());
-    relinkWrappers();
-  } else {
-    std::tie(m_componentInfo, m_detectorInfo) =
-        InstrumentVisitor::makeWrappers(*m_instrument, this);
-  }
-}
-
-/// Sets up links between m_detectorInfo, m_componentInfo, and m_instrument.
-void ParameterMap::relinkWrappers() {
-  m_componentInfo->m_componentInfo->setDetectorInfo(
-      m_detectorInfo->m_detectorInfo.get());
-  m_detectorInfo->m_detectorInfo->setComponentInfo(
-      m_componentInfo->m_componentInfo.get());
-
-  const auto parInstrument = ParComponentFactory::createInstrument(
-      boost::shared_ptr<const Instrument>(m_instrument, NoDeleting()),
-      boost::shared_ptr<ParameterMap>(this, NoDeleting()));
-  m_detectorInfo->m_instrument = parInstrument;
+  std::tie(m_componentInfo, m_detectorInfo) = m_instrument->makeBeamline(*this);
 }
 
 } // Namespace Geometry