Parsing the instrument tree in the base instrument when it is being added to the IDS has two big advantages when loading more than one workspace with the same instrument:
- We avoid repeated parsing of the same tree.
- Data in instruments (`DetectorInfo` and `ComponentInfo`) will be shared between workspaces. Previously this was the case only for workspaces that were created from a parent workspaces, but not indenpendently created workspaces with the same instrument. Note that if there are parameters that modify the instrument sharing will still be (partially) broken. 

Speedup example (second load and beyond) for TOPAZ:
`master`:
```
LoadEmptyInstrument-[Notice] LoadEmptyInstrument successful, Duration 6.08 seconds
LoadEmptyInstrument-[Notice] LoadEmptyInstrument started
LoadEmptyInstrument-[Notice] LoadEmptyInstrument successful, Duration 4.01 seconds
LoadEmptyInstrument-[Notice] LoadEmptyInstrument started
LoadEmptyInstrument-[Notice] LoadEmptyInstrument successful, Duration 3.44 seconds
LoadEmptyInstrument-[Notice] LoadEmptyInstrument started
LoadEmptyInstrument-[Notice] LoadEmptyInstrument successful, Duration 3.40 seconds
LoadEmptyInstrument-[Notice] LoadEmptyInstrument started
LoadEmptyInstrument-[Notice] LoadEmptyInstrument successful, Duration 3.16 seconds
LoadEmptyInstrument-[Notice] LoadEmptyInstrument started
LoadEmptyInstrument-[Notice] LoadEmptyInstrument successful, Duration 3.71 seconds
LoadEmptyInstrument-[Notice] LoadEmptyInstrument started
LoadEmptyInstrument-[Notice] LoadEmptyInstrument successful, Duration 3.58 seconds
```

This branch:
```
LoadEmptyInstrument successful, Duration 5.93 seconds
LoadEmptyInstrument started
LoadEmptyInstrument successful, Duration 2.70 seconds
LoadEmptyInstrument started
LoadEmptyInstrument successful, Duration 1.95 seconds
LoadEmptyInstrument started
LoadEmptyInstrument successful, Duration 2.16 seconds
LoadEmptyInstrument started
LoadEmptyInstrument successful, Duration 1.87 seconds
LoadEmptyInstrument started
LoadEmptyInstrument successful, Duration 2.09 seconds
LoadEmptyInstrument started
LoadEmptyInstrument successful, Duration 1.98 seconds
```


**To test:**

Code review.

Fixes #20163.

Release notes: Probably worth adding this to the release notes, but probably we should just do this once this feature branch is ready for merging into master such that we can summarize all improvements?

---

#### Reviewer ####

Please comment on the following ([full description](http://www.mantidproject.org/Individual_Ticket_Testing)):

##### Code Review #####

- [ ] Is the code of an acceptable quality?
- [ ] Does the code conform to the [coding standards](http://www.mantidproject.org/Coding_Standards)?
- [ ] Are the unit tests small and test the class in isolation?
- [ ] If there are changes in the release notes then do they describe the changes appropriately?

##### Functional Tests #####

- [ ] Do changes function as described? Add comments below that describe the tests performed?
- [ ] Do the changes handle unexpected situations, e.g. bad input?
- [ ] Has the relevant (user and developer) documentation been added/updated?

Does everything look good? Mark the review as **Approve**. A member of `@mantidproject/gatekeepers` will take care of it.