@SimonHeybrock 2017-08-04
@OwenArnold Inspired by one of your comments I moved around things in the init mechanism a fair bit. Let me know what you think, it should either be addressing or obsoleting your comments.