@OwenArnold 2017-08-04
I'm happy with these changes. If nothing flags up with the builds, I'll merge.